refactor: split OpenAPI importer our
